Fix: Refactor image upload, modal content in new component
